Moving model from one app to another without migrations
