+++ suggestions_for_candidates/apps.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+from __future__ import unicode_literals
+
+from django.apps import AppConfig
+
+
+class SuggestionsForCandidatesConfig(AppConfig):
+    name = 'suggestions_for_candidates'
